Type integer appkit version numbers. Only #define appkit version numbers with pre 10.13 SDK

+++ PDFAnnotationCircle_SKExtensions.m
@@ -52,7 +52,7 @@ NSString *SKPDFAnnotationScriptingInteriorColorKey = @"scriptingInteriorColor";
     self = [super initSkimNoteWithBounds:bounds];
     if (self) {
         // PDFAnnotationCircle over-retains the initial PDFBorder ivar on 10.6.x
-        if ((NSInteger)floor(NSAppKitVersionNumber) == NSAppKitVersionNumber10_6)
+        if ((NSInteger)floor(NSAppKitVersionNumber) == (NSInteger)NSAppKitVersionNumber10_6)
             [[self border] release];
         NSColor *color = [[NSUserDefaults standardUserDefaults] colorForKey:SKCircleNoteInteriorColorKey];
         if ([color alphaComponent] > 0.0)
