+++ PDFAnnotationSquare_SKExtensions.m
@@ -52,7 +52,7 @@
     self = [super initSkimNoteWithBounds:bounds];
     if (self) {
         // PDFAnnotationSquare over-retains the initial PDFBorder ivar on 10.6.x
-        if ((NSInteger)floor(NSAppKitVersionNumber) == NSAppKitVersionNumber10_6)
+        if ((NSInteger)floor(NSAppKitVersionNumber) == (NSInteger)NSAppKitVersionNumber10_6)
             [[self border] release];
         NSColor *color = [[NSUserDefaults standardUserDefaults] colorForKey:SKSquareNoteInteriorColorKey];
         if ([color alphaComponent] > 0.0)
