+++ SKNotesPage.m
@@ -53,7 +53,7 @@
 
 - (NSAttributedString *)attributedString {
     // on 10.6 the attributedstring is over-released by one
-    if ((NSInteger)floor(NSAppKitVersionNumber) == NSAppKitVersionNumber10_6)
+    if ((NSInteger)floor(NSAppKitVersionNumber) == (NSInteger)NSAppKitVersionNumber10_6)
         return [[super attributedString] retain];
     return [super attributedString];
 }
